+++ mrb_doc/core/screen_space.rb
@@ -1,13 +1,3 @@
-# Translates the vector to where it is on the screen according to the camera
-# @param vector [Vector2]
-# @param camera [Camera2D]
-# @return [Vector2]
-def get_world_to_screen2D(vector, camera)
-  # mrb_get_world_to_screen2D
-  # src/mruby_integration/core/screen_space.cpp
-  Vector2.new
-end
-
 # Translates the vector to where it is in the world based on it the camera
 # @param vector [Vector2]
 # @param camera [Camera2D]