+++ test/models/camera2d_test.rb
@@ -106,6 +106,13 @@ class Test
 
         assert_equal fixture_camera2d_draw[1], get_screen_data.data
       end
+
+      def test_as_in_viewport
+        camera = Camera2D.new(target: Vector2[2, 1], offset: Vector2[3, 2])
+        vector = Vector2[3, 3]
+
+        assert_equal Vector2[4, 4], camera.as_in_viewport(vector)
+      end
     end
   end
 end